Do not let libmagic try to preserve files' atimes.
In doing so, libmagic ends up changing the files' ctimes.  It
seems the lesser of two evils to allow the atimes to change, and
with many filesystems mounted 'noatime' these days the atimes
don't change anyway.

Futhermore, when libmagic does reinstate the atime it also only
does so to second resolution (even on filesystems with higher
resolutions, such as EXT4).