Avoid logging within signal handling functions.
Calling logit() from within signal handlers is unsafe for two
reasons: logging uses non-async-signal-safe functions, and
arbitrary re-entry can result in a deadlock on the logging mutex.

This commit is an interim solution.  The proper solution is to
create a new signal-handling thread where all signals can be
processed as normal execution code.