Fix delayed client exit after the server exits in some circumstances.
If the server exits while the client is expecting to receive a data
response from it, then the EV_EXIT event gets queued but the expected
EV_DATA event never arrives.  The client will eventually exit with a
receive error when the socket gets closed, but if the socket was
inheritted by a long-running server child process (via On... scripts)
then it will not be closed until the child eventually exits and is
reaped by init(8).

Thanks to: Alex Merenstein